pyproject.toml
[project]
name = "learn-playwright"
version = "0.1.0"
description = "Learn how to use Playwright"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.44.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.1.1",
]

[tool.pytest.ini_options]
testpaths = ["examples"]
asyncio_mode = "auto"
